Reimplement vote on orbSelection, remove duplicate code, abstract utils
