+++ packages/project-disaster-trail/src/components/Game/TaskScreen/TaskMap.js
@@ -1,4 +1,5 @@
-import React from "react";
+/** @jsx jsx */
+import { css, jsx } from "@emotion/core";
 import { PropTypes } from "prop-types";
 import { connect } from "react-redux";
 
@@ -14,6 +15,16 @@ import {
   initialLat
 } from "./mapUtils";
 
+const screenLayout = css`
+  position: relative;
+  overflow: hidden;
+  width: 100%;
+  height: 100%;
+  justify-content: center;
+  align-items: center;
+  background: beige;
+`;
+
 const TaskMap = ({ activeTask, completedTasks, tasks }) => {
   // TODO: Change lon / lat for task
   // const lon = activeTask ? activeTask.locations[0][0] : initialLon;
@@ -21,38 +32,41 @@ const TaskMap = ({ activeTask, completedTasks, tasks }) => {
   const lon = initialLon;
   const lat = initialLat;
 
-  const data = asGeoJSON(tasks, activeTask, completedTasks);
+  const selectedTask = activeTask || tasks[0];
+  const data = asGeoJSON(tasks, selectedTask, completedTasks);
 
   return (
-    <BaseMap
-      initialZoom={15}
-      initialLongitude={lon}
-      initialLatitude={lat}
-      initialPitch={60}
-      navigation={false}
-      useContainerHeight
-      mapGLOptions={{
-        scrollZoom: false,
-        dragPan: false,
-        dragRotate: false,
-        doubleClickZoom: false,
-        touchZoom: false,
-        touchRotate: false,
-        keyboard: false
-      }}
-      animate
-      civicMapStyle="disaster-game"
-    >
-      <IconMap
-        data={data}
-        getPosition={getPosition}
-        iconAtlas="https://i.imgur.com/ZKYqCqW.png"
-        iconMapping={poiIconMapping}
-        iconSizeScale={poiIconZoomScale}
-        getIcon={d => d.properties.type}
-        getSize={() => 12}
-      />
-    </BaseMap>
+    <div css={screenLayout}>
+      <BaseMap
+        initialZoom={15}
+        initialLongitude={lon}
+        initialLatitude={lat}
+        initialPitch={60}
+        navigation={false}
+        useContainerHeight
+        mapGLOptions={{
+          scrollZoom: false,
+          dragPan: false,
+          dragRotate: false,
+          doubleClickZoom: false,
+          touchZoom: false,
+          touchRotate: false,
+          keyboard: false
+        }}
+        animate
+        civicMapStyle="disaster-game"
+      >
+        <IconMap
+          data={data}
+          getPosition={getPosition}
+          iconAtlas="https://i.imgur.com/ZKYqCqW.png"
+          iconMapping={poiIconMapping}
+          iconSizeScale={poiIconZoomScale}
+          getIcon={d => d.properties.type}
+          getSize={() => 12}
+        />
+      </BaseMap>
+    </div>
   );
 };
